+++ database/migrations/2025_04_18_145222_create_payments_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreignId('payer_id')->constrained('users')->onDelete('restrict');
             $table->foreignId('payee_id')->constrained('users')->onDelete('restrict');
             $table->integer('status')->default(PaymentStatusEnum::PENDING);
-            $table->double('value');
+            $table->float('value');
             $table->softDeletes();
             $table->timestamps();
         });
