+++ database/migrations/2025_04_18_202928_create_wallets_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('wallets', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('restrict');
-            $table->double('balance')->default(0);
+            $table->float('balance')->default(0);
             $table->timestamps();
         });
     }
